requirements.txt
Django~=4.0.4
psycopg2-binary==2.9.3
Pillow==9.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
celery==5.2.7
redis==4.3.3


factory-boy==3.2.1
Faker==13.12.0
